// grayjay-android/app/src/unstable/assets/sources/bilibili/src/types.ts
export type Params = {
    readonly [key: string]: string
}
// ...
type HomeFeedItem = {
    readonly goto: "av"
    readonly bvid: string
    readonly cid: number
    readonly uri: string
    readonly pic: string
    readonly title: string
    readonly duration: number
    readonly pubdate: number
    readonly stat: {
        readonly view: number
    }
    readonly owner: HomeFeedItemOwner
} | {
    readonly goto: "ad"
} | {
    readonly goto: "live"
    readonly pic: string
    readonly id: number
    readonly title: string
    readonly room_info: {
        readonly watched_show: {
            readonly num: number
        }
    }
    readonly owner: HomeFeedItemOwner
}

type HomeFeedItemOwner = {
    readonly mid: number
    readonly name: string
    readonly face: string
}

export type HomeFeedResponse = {
    readonly data: {
        readonly item: HomeFeedItem[]
    }
} | null